+++ src/main/java/iuh/fit/se/techgalaxy/service/impl/ProductServiceImpl.java
@@ -61,6 +61,11 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProduct(String id) {
+        long usageCount = productRepository.countOrderDetailsByProductId(id);
+
+        if (usageCount > 0) {
+            throw new IllegalStateException("Cannot delete product as it is referenced in orders.");
+        }
         Product product = productRepository.findById(id).orElseThrow(()-> new RuntimeException("Product not found"));
         productRepository.delete(product);
     }
